# Faster octree nearestKSearch